# contracts/openapi/aura-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aura Core API",
    "version": "1.5.0",
    "description": "Owner-authenticated agent, persona, conversation, run, and memory transport contract. Provider endpoints, credentials, and embedding vectors are never exposed by this API."
  },
  "servers": [{ "url": "/", "description": "Same-origin Aura BFF" }],
  "security": [{ "sessionCookie": [] }],
...
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "operationId": "listModels",
        "tags": ["Models"],
        "responses": {
          "200": { "description": "Current provider-neutral model inventory", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ModelCatalog" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
...
    "/api/v1/memory-model-inventory": {
      "get": {
        "operationId": "getMemoryModelInventory",
        "tags": ["Models", "Memories"],
        "responses": {
          "200": { "description": "Currently verified compatible Ollama model inventory without provider endpoints or credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoryModelInventory" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/v1/memory-model-configuration": {
      "get": {
        "operationId": "getMemoryModelConfiguration",
        "tags": ["Models", "Memories"],
        "responses": {
          "200": { "description": "Current owner memory model configuration and generation state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoryModelConfiguration" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "operationId": "updateMemoryModelConfiguration",
        "tags": ["Models", "Memories"],
        "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateMemoryModelConfigurationRequest" } } } },
        "responses": {
          "200": { "description": "Verified model configuration updated; an embedding change starts or resumes a parallel generation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoryModelConfiguration" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
    "/api/v1/memory-reindex": {
      "get": {
        "operationId": "getMemoryReindexStatus",
        "tags": ["Models", "Memories"],
        "responses": {
          "200": { "description": "Owner-scoped active and replacement embedding generation progress", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoryReindexStatus" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/memory-reindex/resume": {
      "post": {
        "operationId": "resumeMemoryReindex",
        "tags": ["Models", "Memories"],
        "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResumeMemoryReindexRequest" } } } },
        "responses": {
          "202": { "description": "Existing replacement generation reindex queued for resumable processing", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoryReindexStatus" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "__Host-aura_session", "description": "Opaque, rotated, Secure, HttpOnly, SameSite=Lax BFF session cookie." }
    },
    "parameters": {
...
      "CsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string", "minLength": 32, "maxLength": 512 }, "description": "Session-bound anti-CSRF token returned by getSession." },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Caller-generated UUID. Reuse only for an exact replay of the same command." }
    },
    "responses": {
      "BadRequest": { "description": "Malformed request", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Unauthorized": { "description": "No valid Aura session", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Forbidden": { "description": "Authenticated subject is not the configured owner or command validation failed", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "NotFound": { "description": "Resource not found for this principal", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Conflict": { "description": "Version conflict, active run conflict, or idempotency-key payload conflict", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "UnprocessableEntity": { "description": "Semantically invalid command", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "ServiceUnavailable": { "description": "Required dependency unavailable", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    },
    "schemas": {
...
      "ModelCatalog": {
        "type": "object", "additionalProperties": false,
        "required": ["models", "defaultModelId", "observedAt"],
        "properties": {
          "models": { "type": "array", "items": { "$ref": "#/components/schemas/Model" } },
          "defaultModelId": { "type": ["string", "null"] },
          "observedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Model": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "displayName", "provider", "capabilities", "availability", "selectable", "disabledReason"],
        "properties": {
          "id": { "type": "string", "minLength": 1, "maxLength": 255 },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 255 },
          "provider": { "type": "string", "minLength": 1, "maxLength": 64 },
          "capabilities": { "type": "array", "uniqueItems": true, "items": { "type": "string", "enum": ["chat", "completion", "embedding", "vision", "tools", "structured_output"] } },
          "availability": { "type": "string", "enum": ["available", "unavailable", "unknown"] },
          "selectable": { "type": "boolean" },
          "disabledReason": { "type": ["string", "null"], "maxLength": 500 }
        }
      },
...
      "MemoryEmbeddingGenerationStatus": { "type": "string", "enum": ["building", "active", "retired", "failed"] },
...
      "MemoryModelCapability": { "type": "string", "enum": ["structured_output", "embedding"] },
      "MemoryCompatibleModel": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "displayName", "provider", "modelRevision", "modelDigest", "capabilities", "dimension", "available", "disabledReason"],
        "properties": {
          "id": { "type": "string", "minLength": 1, "maxLength": 255 },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 255 },
          "provider": { "type": "string", "minLength": 1, "maxLength": 64 },
          "modelRevision": { "type": ["string", "null"], "maxLength": 255 },
          "modelDigest": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
          "capabilities": { "type": "array", "minItems": 1, "uniqueItems": true, "items": { "$ref": "#/components/schemas/MemoryModelCapability" } },
          "dimension": { "type": ["integer", "null"], "minimum": 1 },
          "available": { "type": "boolean" },
          "disabledReason": { "type": ["string", "null"], "maxLength": 500 }
        }
      },
      "MemoryModelInventory": {
        "type": "object", "additionalProperties": false,
        "required": ["models", "observedAt"],
        "properties": {
          "models": { "type": "array", "items": { "$ref": "#/components/schemas/MemoryCompatibleModel" } },
          "observedAt": { "type": "string", "format": "date-time" }
        }
      },
      "MemoryModelSelection": {
        "type": "object", "additionalProperties": false,
        "required": ["modelId", "modelRevision", "modelDigest"],
        "properties": {
          "modelId": { "type": "string", "minLength": 1, "maxLength": 255 },
          "modelRevision": { "type": ["string", "null"], "maxLength": 255 },
          "modelDigest": { "type": "string", "pattern": "^[0-9a-f]{64}$" }
        }
      },
      "MemoryManagedEmbeddingGeneration": {
        "type": "object", "additionalProperties": false,
        "description": "Owner-scoped generation metadata with exact provider-observed model identity and no vector values.",
        "required": ["id", "generation", "model", "dimension", "status", "createdAt", "activatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "generation": { "type": "integer", "minimum": 1 },
          "model": { "$ref": "#/components/schemas/MemoryModelSelection" },
          "dimension": { "type": "integer", "minimum": 1 },
          "status": { "$ref": "#/components/schemas/MemoryEmbeddingGenerationStatus" },
          "createdAt": { "type": "string", "format": "date-time" },
          "activatedAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "MemoryModelConfiguration": {
        "type": "object", "additionalProperties": false,
        "required": ["version", "extraction", "embedding", "activeGeneration", "buildingGeneration", "updatedAt"],
        "properties": {
          "version": { "type": "integer", "minimum": 1 },
          "extraction": { "$ref": "#/components/schemas/MemoryModelSelection" },
          "embedding": { "$ref": "#/components/schemas/MemoryModelSelection" },
          "activeGeneration": { "oneOf": [{ "$ref": "#/components/schemas/MemoryManagedEmbeddingGeneration" }, { "type": "null" }] },
          "buildingGeneration": { "oneOf": [{ "$ref": "#/components/schemas/MemoryManagedEmbeddingGeneration" }, { "type": "null" }] },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "UpdateMemoryModelConfigurationRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["extractionModelId", "embeddingModelId", "expectedVersion"],
        "properties": {
          "extractionModelId": { "type": "string", "minLength": 1, "maxLength": 255 },
          "embeddingModelId": { "type": "string", "minLength": 1, "maxLength": 255 },
          "expectedVersion": { "type": "integer", "minimum": 1 }
        },
        "description": "Model identity is resolved from current verified inventory by Core; browser-supplied revisions or digests are not accepted."
      },
      "MemoryReindexPhase": { "type": "string", "enum": ["idle", "queued", "running", "failed", "completed"] },
      "MemoryReindexStatus": {
        "type": "object", "additionalProperties": false,
        "required": ["phase", "activeGeneration", "replacementGeneration", "processedRevisionCount", "totalRevisionCount", "startedAt", "updatedAt", "completedAt", "retryable"],
        "properties": {
          "phase": { "$ref": "#/components/schemas/MemoryReindexPhase" },
          "activeGeneration": { "oneOf": [{ "$ref": "#/components/schemas/MemoryManagedEmbeddingGeneration" }, { "type": "null" }] },
          "replacementGeneration": { "oneOf": [{ "$ref": "#/components/schemas/MemoryManagedEmbeddingGeneration" }, { "type": "null" }] },
          "processedRevisionCount": { "type": "integer", "minimum": 0 },
          "totalRevisionCount": { "type": "integer", "minimum": 0 },
          "startedAt": { "type": ["string", "null"], "format": "date-time" },
          "updatedAt": { "type": ["string", "null"], "format": "date-time" },
          "completedAt": { "type": ["string", "null"], "format": "date-time" },
          "retryable": { "type": "boolean" }
        }
      },
      "ResumeMemoryReindexRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["generationId"],
        "properties": { "generationId": { "type": "string", "format": "uuid" } }
      },
...
      "ErrorEnvelope": {
        "type": "object", "additionalProperties": false,
        "required": ["type", "title", "status", "code", "detail", "retryable", "traceId", "resource"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_]{2,63}$" },
          "detail": { "type": "string", "minLength": 1, "maxLength": 1000 },
          "retryable": { "type": "boolean" },
          "traceId": { "type": "string", "pattern": "^[0-9a-f]{32}$" },
          "resource": { "type": ["string", "null"], "maxLength": 255 },
          "fieldErrors": { "type": "array", "items": { "$ref": "#/components/schemas/FieldError" } }
        }
      },
      "FieldError": {
        "type": "object", "additionalProperties": false,
        "required": ["field", "code", "message"],
        "properties": { "field": { "type": "string", "minLength": 1, "maxLength": 255 }, "code": { "type": "string", "minLength": 1, "maxLength": 64 }, "message": { "type": "string", "minLength": 1, "maxLength": 500 } }
      },
...
    }
  }
}